+++ bpf-metrics-exporter/src/main.rs
@@ -1,5 +1,5 @@
 use aya::{loaded_programs, maps::loaded_maps, programs::loaded_links};
-use bpfman_api::ProgramType;
+use bpfman::types::ProgramType;
 use chrono::{prelude::DateTime, Utc};
 use clap::Parser;
 use opentelemetry::{
