+++ bpfman/src/bin/cli/table.rs
@@ -1,23 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of bpfman
 
-use anyhow::bail;
-use bpfman_api::{
-    v1::{
-        attach_info::Info, bytecode_location::Location, list_response::ListResult,
-        FentryAttachInfo, FexitAttachInfo, KernelProgramInfo, KprobeAttachInfo, ProgramInfo,
-        TcAttachInfo, TracepointAttachInfo, UprobeAttachInfo, XdpAttachInfo,
-    },
-    ImagePullPolicy,
-    ProbeType::{Kprobe, Kretprobe, Uprobe, Uretprobe},
-    ProgramType, TcProceedOn, XdpProceedOn,
-};
+use bpfman::types::{ImagePullPolicy, Location, ProbeType::*, Program};
 use comfy_table::{Cell, Color, Table};
 use hex::encode_upper;
 pub(crate) struct ProgTable(Table);
 
 impl ProgTable {
-    pub(crate) fn new_get_bpfman(r: &Option<ProgramInfo>) -> Result<Self, anyhow::Error> {
+    pub(crate) fn new_program(program: &Program) -> Result<Self, anyhow::Error> {
         let mut table = Table::new();
 
         table.load_preset(comfy_table::presets::NOTHING);
@@ -26,45 +16,34 @@ impl ProgTable {
             .add_attribute(comfy_table::Attribute::Underlined)
             .fg(Color::Green)]);
 
-        if r.is_none() {
-            table.add_row(vec!["NONE"]);
-            return Ok(ProgTable(table));
-        }
-        let info = r.clone().unwrap();
+        let data = program.get_data();
 
-        if info.bytecode.is_none() {
-            table.add_row(vec!["NONE"]);
-            return Ok(ProgTable(table));
-        }
-
-        if info.name.clone().is_empty() {
+        let name = data.get_name()?;
+        if name.is_empty() {
             table.add_row(vec!["Name:", "None"]);
         } else {
-            table.add_row(vec!["Name:", &info.name.clone()]);
+            table.add_row(vec!["Name:", &name]);
         }
 
-        match info.bytecode.clone().unwrap().location.clone() {
-            Some(l) => match l {
-                Location::Image(i) => {
-                    table.add_row(vec!["Image URL:", &i.url]);
-                    table.add_row(vec!["Pull Policy:", &format!{ "{}", TryInto::<ImagePullPolicy>::try_into(i.image_pull_policy)?}]);
-                }
-                Location::File(p) => {
-                    table.add_row(vec!["Path:", &p]);
-                }
-            },
-            // not a bpfman program
-            None => {
-                table.add_row(vec!["NONE"]);
-                return Ok(ProgTable(table));
+        match data.get_location()? {
+            Location::Image(i) => {
+                table.add_row(vec!["Image URL:", &i.image_url]);
+                table.add_row(vec![
+                    "Pull Policy:",
+                    &format! { "{}", TryInto::<ImagePullPolicy>::try_into(i.image_pull_policy)?},
+                ]);
             }
-        }
+            Location::File(p) => {
+                table.add_row(vec!["Path:", &p]);
+            }
+        };
 
-        if info.global_data.is_empty() {
+        let global_data = data.get_global_data()?;
+        if global_data.is_empty() {
             table.add_row(vec!["Global:", "None"]);
         } else {
             let mut first = true;
-            for (key, value) in info.global_data.clone() {
+            for (key, value) in global_data {
                 let data = &format! {"{key}={}", encode_upper(value)};
                 if first {
                     first = false;
@@ -75,11 +54,12 @@ impl ProgTable {
             }
         }
 
-        if info.metadata.is_empty() {
+        let metadata = data.get_metadata()?;
+        if metadata.is_empty() {
             table.add_row(vec!["Metadata:", "None"]);
         } else {
             let mut first = true;
-            for (key, value) in info.metadata.clone() {
+            for (key, value) in metadata.clone() {
                 let data = &format! {"{key}={value}"};
                 if first {
                     first = false;
@@ -90,127 +70,112 @@ impl ProgTable {
             }
         }
 
-        if info.map_pin_path.clone().is_empty() {
-            table.add_row(vec!["Map Pin Path:", "None"]);
+        if let Some(map_pin_path) = data.get_map_pin_path()? {
+            table.add_row(vec![
+                "Map Pin Path:",
+                map_pin_path
+                    .to_str()
+                    .expect("map_pin_path is not valid Unicode"),
+            ]);
         } else {
-            table.add_row(vec!["Map Pin Path:", &info.map_pin_path.clone()]);
+            table.add_row(vec!["Map Pin Path:", "None"]);
         }
 
-        match info.map_owner_id {
+        match data.get_map_owner_id()? {
             Some(id) => table.add_row(vec!["Map Owner ID:", &id.to_string()]),
             None => table.add_row(vec!["Map Owner ID:", "None"]),
         };
 
-        if info.map_used_by.clone().is_empty() {
+        let map_used_by = data.get_maps_used_by()?;
+        if map_used_by.is_empty() {
             table.add_row(vec!["Maps Used By:", "None"]);
         } else {
             let mut first = true;
-            for prog_id in info.clone().map_used_by {
+            for prog_id in map_used_by {
                 if first {
                     first = false;
-                    table.add_row(vec!["Maps Used By:", &prog_id]);
+                    table.add_row(vec!["Maps Used By:", &prog_id.to_string()]);
                 } else {
-                    table.add_row(vec!["", &prog_id]);
+                    table.add_row(vec!["", &prog_id.to_string()]);
                 }
             }
         };
 
-        if info.attach.is_some() {
-            match info.attach.clone().unwrap().info.unwrap() {
-                Info::XdpAttachInfo(XdpAttachInfo {
-                    priority,
-                    iface,
-                    position,
-                    proceed_on,
-                }) => {
-                    let proc_on = match XdpProceedOn::from_int32s(proceed_on) {
-                        Ok(p) => p,
-                        Err(e) => bail!("error parsing proceed_on {e}"),
-                    };
+        match program {
+            Program::Xdp(p) => {
+                table.add_row(vec!["Priority:", &p.get_priority()?.to_string()]);
+                table.add_row(vec!["Iface:", &p.get_iface()?]);
+                table.add_row(vec![
+                    "Position:",
+                    &match p.get_current_position()? {
+                        Some(pos) => pos.to_string(),
+                        None => "NONE".to_string(),
+                    },
+                ]);
+                table.add_row(vec!["Proceed On:", &format!("{}", p.get_proceed_on()?)]);
+            }
+            Program::Tc(p) => {
+                table.add_row(vec!["Priority:", &p.get_priority()?.to_string()]);
+                table.add_row(vec!["Iface:", &p.get_iface()?]);
+                table.add_row(vec![
+                    "Position:",
+                    &match p.get_current_position()? {
+                        Some(pos) => pos.to_string(),
+                        None => "NONE".to_string(),
+                    },
+                ]);
+                table.add_row(vec!["Direction:", &p.get_direction()?.to_string()]);
+                table.add_row(vec!["Proceed On:", &format!("{}", p.get_proceed_on()?)]);
+            }
+            Program::Tracepoint(p) => {
+                table.add_row(vec!["Tracepoint:", &p.get_tracepoint()?]);
+            }
+            Program::Kprobe(p) => {
+                let probe_type = match p.get_retprobe()? {
+                    true => Kretprobe,
+                    false => Kprobe,
+                };
 
-                    table.add_row(vec!["Priority:", &priority.to_string()]);
-                    table.add_row(vec!["Iface:", &iface]);
-                    table.add_row(vec!["Position:", &position.to_string()]);
-                    table.add_row(vec!["Proceed On:", &format!("{proc_on}")]);
-                }
-                Info::TcAttachInfo(TcAttachInfo {
-                    priority,
-                    iface,
-                    position,
-                    direction,
-                    proceed_on,
-                }) => {
-                    let proc_on = match TcProceedOn::from_int32s(proceed_on) {
-                        Ok(p) => p,
-                        Err(e) => bail!("error parsing proceed_on {e}"),
-                    };
-
-                    table.add_row(vec!["Priority:", &priority.to_string()]);
-                    table.add_row(vec!["Iface:", &iface]);
-                    table.add_row(vec!["Position:", &position.to_string()]);
-                    table.add_row(vec!["Direction:", &direction]);
-                    table.add_row(vec!["Proceed On:", &format!("{proc_on}")]);
-                }
-                Info::TracepointAttachInfo(TracepointAttachInfo { tracepoint }) => {
-                    table.add_row(vec!["Tracepoint:", &tracepoint]);
-                }
-                Info::KprobeAttachInfo(KprobeAttachInfo {
-                    fn_name,
-                    offset,
-                    retprobe,
-                    container_pid,
-                }) => {
-                    let probe_type = match retprobe {
-                        true => Kretprobe,
-                        false => Kprobe,
-                    };
-
-                    table.add_row(vec!["Probe Type:", &format!["{probe_type}"]]);
-                    table.add_row(vec!["Function Name:", &fn_name]);
-                    table.add_row(vec!["Offset:", &offset.to_string()]);
-                    table.add_row(vec![
-                        "Container PID",
-                        &container_pid.unwrap_or(0).to_string(),
-                    ]);
-                }
-                Info::UprobeAttachInfo(UprobeAttachInfo {
-                    fn_name,
-                    offset,
-                    target,
-                    retprobe,
-                    pid,
-                    container_pid,
-                }) => {
-                    let probe_type = match retprobe {
-                        true => Uretprobe,
-                        false => Uprobe,
-                    };
-
-                    table.add_row(vec!["Probe Type:", &format!["{probe_type}"]]);
-                    table.add_row(vec!["Function Name:", &fn_name.unwrap_or("".to_string())]);
-                    table.add_row(vec!["Offset:", &offset.to_string()]);
-                    table.add_row(vec!["Target:", &target]);
-                    table.add_row(vec!["PID", &pid.unwrap_or(0).to_string()]);
-                    table.add_row(vec![
-                        "Container PID",
-                        &container_pid.unwrap_or(0).to_string(),
-                    ]);
-                }
-                Info::FentryAttachInfo(FentryAttachInfo { fn_name }) => {
-                    table.add_row(vec!["Function Name:", &fn_name]);
-                }
-                Info::FexitAttachInfo(FexitAttachInfo { fn_name }) => {
-                    table.add_row(vec!["Function Name:", &fn_name]);
-                }
+                table.add_row(vec!["Probe Type:", &format!["{probe_type}"]]);
+                table.add_row(vec!["Function Name:", &p.get_fn_name()?]);
+                table.add_row(vec!["Offset:", &p.get_offset()?.to_string()]);
+                table.add_row(vec![
+                    "PID:",
+                    &p.get_container_pid()?.unwrap_or(0).to_string(),
+                ]);
+            }
+            Program::Uprobe(p) => {
+                let probe_type = match p.get_retprobe()? {
+                    true => Kretprobe,
+                    false => Kprobe,
+                };
+                table.add_row(vec!["Probe Type:", &format!["{probe_type}"]]);
+                table.add_row(vec![
+                    "Function Name:",
+                    &p.get_fn_name()?.unwrap_or("NONE".to_string()),
+                ]);
+                table.add_row(vec!["Offset:", &p.get_offset()?.to_string()]);
+                table.add_row(vec!["Target:", &p.get_target()?]);
+                table.add_row(vec!["PID", &p.get_pid()?.unwrap_or(0).to_string()]);
+                table.add_row(vec![
+                    "Container PID:",
+                    &p.get_container_pid()?.unwrap_or(0).to_string(),
+                ]);
+            }
+            Program::Fentry(p) => {
+                table.add_row(vec!["Function Name:", &p.get_fn_name()?]);
+            }
+            Program::Fexit(p) => {
+                table.add_row(vec!["Function Name:", &p.get_fn_name()?]);
+            }
+            Program::Unsupported(_) => {
+                table.add_row(vec!["Unsupported Program Type", "None"]);
             }
         }
-
         Ok(ProgTable(table))
     }
 
-    pub(crate) fn new_get_unsupported(
-        r: &Option<KernelProgramInfo>,
-    ) -> Result<Self, anyhow::Error> {
+    pub(crate) fn new_kernel_info(r: &Program) -> Result<Self, anyhow::Error> {
         let mut table = Table::new();
 
         table.load_preset(comfy_table::presets::NOTHING);
@@ -219,53 +184,47 @@ impl ProgTable {
             .add_attribute(comfy_table::Attribute::Underlined)
             .fg(Color::Green)]);
 
-        if r.is_none() {
-            table.add_row(vec!["NONE"]);
-            return Ok(ProgTable(table));
-        }
-        let kernel_info = r.clone().unwrap();
-
-        let name = if kernel_info.name.clone().is_empty() {
+        let p = r.get_data();
+        let name = if p.get_kernel_name()?.is_empty() {
             "None".to_string()
         } else {
-            kernel_info.name.clone()
+            p.get_kernel_name()?
         };
 
         let rows = vec![
-            vec!["ID:".to_string(), kernel_info.id.to_string()],
+            vec!["ID:".to_string(), p.get_id()?.to_string()],
             vec!["Name:".to_string(), name],
-            vec![
-                "Type:".to_string(),
-                format!("{}", ProgramType::try_from(kernel_info.program_type)?),
-            ],
-            vec!["Loaded At:".to_string(), kernel_info.loaded_at.clone()],
-            vec!["Tag:".to_string(), kernel_info.tag.clone()],
+            vec!["Type:".to_string(), format!("{}", r.kind())],
+            vec!["Loaded At:".to_string(), p.get_kernel_loaded_at()?],
+            vec!["Tag:".to_string(), p.get_kernel_tag()?],
             vec![
                 "GPL Compatible:".to_string(),
-                kernel_info.gpl_compatible.to_string(),
+                p.get_kernel_gpl_compatible()?.to_string(),
             ],
-            vec!["Map IDs:".to_string(), format!("{:?}", kernel_info.map_ids)],
-            vec!["BTF ID:".to_string(), kernel_info.btf_id.to_string()],
+            vec![
+                "Map IDs:".to_string(),
+                format!("{:?}", p.get_kernel_map_ids()?),
+            ],
+            vec!["BTF ID:".to_string(), p.get_kernel_btf_id()?.to_string()],
             vec![
                 "Size Translated (bytes):".to_string(),
-                kernel_info.bytes_xlated.to_string(),
+                p.get_kernel_bytes_xlated()?.to_string(),
             ],
-            vec!["JITted:".to_string(), kernel_info.jited.to_string()],
+            vec!["JITted:".to_string(), p.get_kernel_jited()?.to_string()],
             vec![
                 "Size JITted:".to_string(),
-                kernel_info.bytes_jited.to_string(),
+                p.get_kernel_bytes_jited()?.to_string(),
             ],
             vec![
                 "Kernel Allocated Memory (bytes):".to_string(),
-                kernel_info.bytes_memlock.to_string(),
+                p.get_kernel_bytes_memlock()?.to_string(),
             ],
             vec![
                 "Verified Instruction Count:".to_string(),
-                kernel_info.verified_insns.to_string(),
+                p.get_kernel_verified_insns()?.to_string(),
             ],
         ];
         table.add_rows(rows);
-
         Ok(ProgTable(table))
     }
 
@@ -287,18 +246,14 @@ impl ProgTable {
         self.0.add_row(vec![id, name, type_, load_time]);
     }
 
-    pub(crate) fn add_response_prog(&mut self, r: ListResult) -> anyhow::Result<()> {
-        if r.kernel_info.is_none() {
-            self.0.add_row(vec!["NONE"]);
-            return Ok(());
-        }
-        let kernel_info = r.kernel_info.unwrap();
+    pub(crate) fn add_response_prog(&mut self, r: Program) -> anyhow::Result<()> {
+        let data = r.get_data();
 
         self.add_row_list(
-            kernel_info.id.to_string(),
-            kernel_info.name,
-            (ProgramType::try_from(kernel_info.program_type)?).to_string(),
-            kernel_info.loaded_at,
+            data.get_id()?.to_string(),
+            data.get_kernel_name()?,
+            r.kind().to_string(),
+            data.get_kernel_loaded_at()?,
         );
 
         Ok(())
