+++ bpfman/src/types.rs
@@ -10,31 +10,23 @@ use std::{
 };
 
 use aya::programs::ProgramInfo as AyaProgInfo;
-use bpfman_api::{
-    util::directories::RTDIR_FS,
-    v1::{
-        attach_info::Info, bytecode_location::Location as V1Location, AttachInfo, BytecodeLocation,
-        FentryAttachInfo, FexitAttachInfo, KernelProgramInfo as V1KernelProgramInfo,
-        KprobeAttachInfo, ProgramInfo as V1ProgramInfo, TcAttachInfo, TracepointAttachInfo,
-        UprobeAttachInfo, XdpAttachInfo,
-    },
-    ParseError, ProgramType, TcProceedOn, TcProceedOnEntry, XdpProceedOn, XdpProceedOnEntry,
-};
 use chrono::{prelude::DateTime, Local};
+use clap::ValueEnum;
 use log::{info, warn};
 use rand::Rng;
 use serde::{Deserialize, Serialize};
+use sled::Db;
 use tokio::sync::oneshot;
 
 use crate::{
-    errors::BpfmanError,
+    directories::RTDIR_FS,
+    errors::{BpfmanError, ParseError},
     multiprog::{DispatcherId, DispatcherInfo},
     oci_utils::image_manager::{BytecodeImage, ImageManager},
     utils::{
         bytes_to_bool, bytes_to_i32, bytes_to_string, bytes_to_u32, bytes_to_u64, bytes_to_usize,
         sled_get, sled_get_option, sled_insert,
     },
-    ROOT_DB,
 };
 
 /// These constants define the key of SLED DB
@@ -98,26 +90,11 @@ const UPROBE_PID: &str = "uprobe_pid";
 const UPROBE_TARGET: &str = "uprobe_target";
 
 const FENTRY_FN_NAME: &str = "fentry_fn_name";
-
 const FEXIT_FN_NAME: &str = "fexit_fn_name";
 
 /// Provided by the requester and used by the manager task to send
 /// the command response back to the requester.
 type Responder<T> = oneshot::Sender<T>;
-
-/// Multiple different commands are multiplexed over a single channel.
-#[derive(Debug)]
-pub enum Command {
-    /// Load a program
-    Load(LoadArgs),
-    Unload(UnloadArgs),
-    List {
-        responder: Responder<Vec<Program>>,
-        filter: ListFilter,
-    },
-    Get(GetArgs),
-    PullBytecode(PullBytecodeArgs),
-}
 
 #[derive(Debug, Clone, Default)]
 pub struct ListFilter {
@@ -198,8 +175,8 @@ impl ListFilter {
 
 #[derive(Debug)]
 pub struct LoadArgs {
-    pub(crate) program: Program,
-    pub(crate) responder: Responder<Result<Program, BpfmanError>>,
+    pub program: Program,
+    pub responder: Responder<Result<Program, BpfmanError>>,
 }
 
 #[derive(Debug, Clone)]
@@ -214,24 +191,6 @@ pub enum Program {
     Unsupported(ProgramData),
 }
 
-#[derive(Debug)]
-pub struct UnloadArgs {
-    pub(crate) id: u32,
-    pub(crate) responder: Responder<Result<(), BpfmanError>>,
-}
-
-#[derive(Debug)]
-pub struct GetArgs {
-    pub(crate) id: u32,
-    pub(crate) responder: Responder<Result<Program, BpfmanError>>,
-}
-
-#[derive(Debug)]
-pub struct PullBytecodeArgs {
-    pub(crate) image: BytecodeImage,
-    pub(crate) responder: Responder<Result<(), BpfmanError>>,
-}
-
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub enum Location {
     Image(BytecodeImage),
@@ -241,6 +200,7 @@ pub enum Location {
 impl Location {
     async fn get_program_bytes(
         &self,
+        root_db: &Db,
         image_manager: &mut ImageManager,
     ) -> Result<(Vec<u8>, String), BpfmanError> {
         match self {
@@ -248,13 +208,14 @@ impl Location {
             Location::Image(l) => {
                 let (path, bpf_function_name) = image_manager
                     .get_image(
+                        root_db,
                         &l.image_url,
                         l.image_pull_policy.clone(),
                         l.username.clone(),
                         l.password.clone(),
                     )
                     .await?;
-                let bytecode = image_manager.get_bytecode_from_image_store(path)?;
+                let bytecode = image_manager.get_bytecode_from_image_store(root_db, path)?;
 
                 Ok((bytecode, bpf_function_name))
             }
@@ -305,118 +266,7 @@ impl std::fmt::Display for Direction {
     }
 }
 
-impl TryFrom<&Program> for V1ProgramInfo {
-    type Error = BpfmanError;
-
-    fn try_from(program: &Program) -> Result<Self, Self::Error> {
-        let data: &ProgramData = program.get_data();
-
-        let bytecode = match program.location()? {
-            crate::command::Location::Image(m) => {
-                Some(BytecodeLocation {
-                    location: Some(V1Location::Image(bpfman_api::v1::BytecodeImage {
-                        url: m.get_url().to_string(),
-                        image_pull_policy: m.get_pull_policy().to_owned() as i32,
-                        // Never dump Plaintext Credentials
-                        username: Some(String::new()),
-                        password: Some(String::new()),
-                    })),
-                })
-            }
-            crate::command::Location::File(m) => Some(BytecodeLocation {
-                location: Some(V1Location::File(m.to_string())),
-            }),
-        };
-
-        let attach_info = AttachInfo {
-            info: match program.clone() {
-                Program::Xdp(p) => Some(Info::XdpAttachInfo(XdpAttachInfo {
-                    priority: p.get_priority()?,
-                    iface: p.get_iface()?.to_string(),
-                    position: p.get_current_position()?.unwrap_or(0) as i32,
-                    proceed_on: p.get_proceed_on()?.as_action_vec(),
-                })),
-                Program::Tc(p) => Some(Info::TcAttachInfo(TcAttachInfo {
-                    priority: p.get_priority()?,
-                    iface: p.get_iface()?.to_string(),
-                    position: p.get_current_position()?.unwrap_or(0) as i32,
-                    direction: p.get_direction()?.to_string(),
-                    proceed_on: p.get_proceed_on()?.as_action_vec(),
-                })),
-                Program::Tracepoint(p) => Some(Info::TracepointAttachInfo(TracepointAttachInfo {
-                    tracepoint: p.get_tracepoint()?.to_string(),
-                })),
-                Program::Kprobe(p) => Some(Info::KprobeAttachInfo(KprobeAttachInfo {
-                    fn_name: p.get_fn_name()?.to_string(),
-                    offset: p.get_offset()?,
-                    retprobe: p.get_retprobe()?,
-                    container_pid: p.get_container_pid()?,
-                })),
-                Program::Uprobe(p) => Some(Info::UprobeAttachInfo(UprobeAttachInfo {
-                    fn_name: p.get_fn_name()?.map(|v| v.to_string()),
-                    offset: p.get_offset()?,
-                    target: p.get_target()?.to_string(),
-                    retprobe: p.get_retprobe()?,
-                    pid: p.get_pid()?,
-                    container_pid: p.get_container_pid()?,
-                })),
-                Program::Fentry(p) => Some(Info::FentryAttachInfo(FentryAttachInfo {
-                    fn_name: p.get_fn_name()?.to_string(),
-                })),
-                Program::Fexit(p) => Some(Info::FexitAttachInfo(FexitAttachInfo {
-                    fn_name: p.get_fn_name()?.to_string(),
-                })),
-                Program::Unsupported(_) => None,
-            },
-        };
-
-        // Populate the Program Info with bpfman data
-        Ok(V1ProgramInfo {
-            name: data.get_name()?.to_string(),
-            bytecode,
-            attach: Some(attach_info),
-            global_data: data.get_global_data()?,
-            map_owner_id: data.get_map_owner_id()?,
-            map_pin_path: data
-                .get_map_pin_path()?
-                .map_or(String::new(), |v| v.to_str().unwrap().to_string()),
-            map_used_by: data
-                .get_maps_used_by()?
-                .iter()
-                .map(|m| m.to_string())
-                .collect(),
-            metadata: data.get_metadata()?,
-        })
-    }
-}
-
-impl TryFrom<&Program> for V1KernelProgramInfo {
-    type Error = BpfmanError;
-
-    fn try_from(program: &Program) -> Result<Self, Self::Error> {
-        // Get the Kernel Info.
-        let data: &ProgramData = program.get_data();
-
-        // Populate the Kernel Info.
-        Ok(V1KernelProgramInfo {
-            id: data.get_id()?,
-            name: data.get_kernel_name()?.to_string(),
-            program_type: program.kind() as u32,
-            loaded_at: data.get_kernel_loaded_at()?.to_string(),
-            tag: data.get_kernel_tag()?.to_string(),
-            gpl_compatible: data.get_kernel_gpl_compatible()?,
-            map_ids: data.get_kernel_map_ids()?,
-            btf_id: data.get_kernel_btf_id()?,
-            bytes_xlated: data.get_kernel_bytes_xlated()?,
-            jited: data.get_kernel_jited()?,
-            bytes_jited: data.get_kernel_bytes_jited()?,
-            bytes_memlock: data.get_kernel_bytes_memlock()?,
-            verified_insns: data.get_kernel_verified_insns()?,
-        })
-    }
-}
-
-/// ProgramInfo stores information about bpf programs that are loaded and managed
+/// ProgramData stores information about bpf programs that are loaded and managed
 /// by bpfman.
 #[derive(Debug, Clone)]
 pub struct ProgramData {
@@ -430,6 +280,7 @@ impl ProgramData {
         Self { db_tree: tree }
     }
     pub fn new_pre_load(
+        root_db: &Db,
         location: Location,
         name: String,
         metadata: HashMap<String, String>,
@@ -439,7 +290,7 @@ impl ProgramData {
         let mut rng = rand::thread_rng();
         let id_rand = rng.gen::<u32>();
 
-        let db_tree = ROOT_DB
+        let db_tree = root_db
             .open_tree(PROGRAM_PRE_LOAD_PREFIX.to_string() + &id_rand.to_string())
             .expect("Unable to open program database tree");
 
@@ -457,8 +308,8 @@ impl ProgramData {
         Ok(pd)
     }
 
-    pub(crate) fn swap_tree(&mut self, new_id: u32) -> Result<(), BpfmanError> {
-        let new_tree = ROOT_DB
+    pub(crate) fn swap_tree(&mut self, root_db: &Db, new_id: u32) -> Result<(), BpfmanError> {
+        let new_tree = root_db
             .open_tree(PROGRAM_PREFIX.to_string() + &new_id.to_string())
             .expect("Unable to open program database tree");
 
@@ -473,7 +324,7 @@ impl ProgramData {
             })?;
         }
 
-        ROOT_DB
+        root_db
             .drop_tree(self.db_tree.name())
             .expect("unable to delete temporary program tree");
 
@@ -500,7 +351,7 @@ impl ProgramData {
         )
     }
 
-    fn get_kind(&self) -> Result<Option<ProgramType>, BpfmanError> {
+    pub fn get_kind(&self) -> Result<Option<ProgramType>, BpfmanError> {
         sled_get_option(&self.db_tree, KIND).map(|v| v.map(|v| bytes_to_u32(v).try_into().unwrap()))
     }
 
@@ -508,7 +359,7 @@ impl ProgramData {
         sled_insert(&self.db_tree, NAME, name.as_bytes())
     }
 
-    pub(crate) fn get_name(&self) -> Result<String, BpfmanError> {
+    pub fn get_name(&self) -> Result<String, BpfmanError> {
         sled_get(&self.db_tree, NAME).map(|v| bytes_to_string(&v))
     }
 
@@ -516,7 +367,7 @@ impl ProgramData {
         sled_insert(&self.db_tree, ID, &id.to_ne_bytes())
     }
 
-    pub(crate) fn get_id(&self) -> Result<u32, BpfmanError> {
+    pub fn get_id(&self) -> Result<u32, BpfmanError> {
         sled_get(&self.db_tree, ID).map(bytes_to_u32)
     }
 
@@ -551,7 +402,7 @@ impl ProgramData {
         })
     }
 
-    pub(crate) fn get_location(&self) -> Result<Location, BpfmanError> {
+    pub fn get_location(&self) -> Result<Location, BpfmanError> {
         if let Ok(l) = sled_get(&self.db_tree, LOCATION_FILENAME) {
             Ok(Location::File(bytes_to_string(&l).to_string()))
         } else {
@@ -586,7 +437,7 @@ impl ProgramData {
         })
     }
 
-    pub(crate) fn get_global_data(&self) -> Result<HashMap<String, Vec<u8>>, BpfmanError> {
+    pub fn get_global_data(&self) -> Result<HashMap<String, Vec<u8>>, BpfmanError> {
         self.db_tree
             .scan_prefix(PREFIX_GLOBAL_DATA)
             .map(|n| {
@@ -624,7 +475,7 @@ impl ProgramData {
         })
     }
 
-    pub(crate) fn get_metadata(&self) -> Result<HashMap<String, String>, BpfmanError> {
+    pub fn get_metadata(&self) -> Result<HashMap<String, String>, BpfmanError> {
         self.db_tree
             .scan_prefix(PREFIX_METADATA)
             .map(|n| {
@@ -650,7 +501,7 @@ impl ProgramData {
         sled_insert(&self.db_tree, MAP_OWNER_ID, &id.to_ne_bytes())
     }
 
-    pub(crate) fn get_map_owner_id(&self) -> Result<Option<u32>, BpfmanError> {
+    pub fn get_map_owner_id(&self) -> Result<Option<u32>, BpfmanError> {
         sled_get_option(&self.db_tree, MAP_OWNER_ID).map(|v| v.map(bytes_to_u32))
     }
 
@@ -662,7 +513,7 @@ impl ProgramData {
         )
     }
 
-    pub(crate) fn get_map_pin_path(&self) -> Result<Option<PathBuf>, BpfmanError> {
+    pub fn get_map_pin_path(&self) -> Result<Option<PathBuf>, BpfmanError> {
         sled_get_option(&self.db_tree, MAP_PIN_PATH)
             .map(|v| v.map(|f| PathBuf::from(bytes_to_string(&f))))
     }
@@ -680,7 +531,7 @@ impl ProgramData {
         })
     }
 
-    pub(crate) fn get_maps_used_by(&self) -> Result<Vec<u32>, BpfmanError> {
+    pub fn get_maps_used_by(&self) -> Result<Vec<u32>, BpfmanError> {
         self.db_tree
             .scan_prefix(PREFIX_MAPS_USED_BY)
             .map(|n| n.map(|(_, v)| bytes_to_u32(v.to_vec())))
@@ -709,10 +560,11 @@ impl ProgramData {
 
     pub(crate) async fn set_program_bytes(
         &mut self,
+        root_db: &Db,
         image_manager: &mut ImageManager,
     ) -> Result<(), BpfmanError> {
         let loc = self.get_location()?;
-        match loc.get_program_bytes(image_manager).await {
+        match loc.get_program_bytes(root_db, image_manager).await {
             Err(e) => Err(e),
             Ok((v, s)) => {
                 match loc {
@@ -753,7 +605,7 @@ impl ProgramData {
      * Methods for setting and getting kernel information.
      */
 
-    pub(crate) fn get_kernel_name(&self) -> Result<String, BpfmanError> {
+    pub fn get_kernel_name(&self) -> Result<String, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_NAME).map(|n| bytes_to_string(&n))
     }
 
@@ -761,7 +613,7 @@ impl ProgramData {
         sled_insert(&self.db_tree, KERNEL_NAME, name.as_bytes())
     }
 
-    pub(crate) fn get_kernel_program_type(&self) -> Result<u32, BpfmanError> {
+    pub fn get_kernel_program_type(&self) -> Result<u32, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_PROGRAM_TYPE).map(bytes_to_u32)
     }
 
@@ -773,7 +625,7 @@ impl ProgramData {
         )
     }
 
-    pub(crate) fn get_kernel_loaded_at(&self) -> Result<String, BpfmanError> {
+    pub fn get_kernel_loaded_at(&self) -> Result<String, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_LOADED_AT).map(|n| bytes_to_string(&n))
     }
 
@@ -791,7 +643,7 @@ impl ProgramData {
         )
     }
 
-    pub(crate) fn get_kernel_tag(&self) -> Result<String, BpfmanError> {
+    pub fn get_kernel_tag(&self) -> Result<String, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_TAG).map(|n| bytes_to_string(&n))
     }
 
@@ -814,11 +666,11 @@ impl ProgramData {
         )
     }
 
-    pub(crate) fn get_kernel_gpl_compatible(&self) -> Result<bool, BpfmanError> {
+    pub fn get_kernel_gpl_compatible(&self) -> Result<bool, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_GPL_COMPATIBLE).map(bytes_to_bool)
     }
 
-    pub(crate) fn get_kernel_map_ids(&self) -> Result<Vec<u32>, BpfmanError> {
+    pub fn get_kernel_map_ids(&self) -> Result<Vec<u32>, BpfmanError> {
         self.db_tree
             .scan_prefix(PREFIX_KERNEL_MAP_IDS.as_bytes())
             .map(|n| n.map(|(_, v)| bytes_to_u32(v.to_vec())))
@@ -842,7 +694,7 @@ impl ProgramData {
         })
     }
 
-    pub(crate) fn get_kernel_btf_id(&self) -> Result<u32, BpfmanError> {
+    pub fn get_kernel_btf_id(&self) -> Result<u32, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_BTF_ID).map(bytes_to_u32)
     }
 
@@ -850,7 +702,7 @@ impl ProgramData {
         sled_insert(&self.db_tree, KERNEL_BTF_ID, &btf_id.to_ne_bytes())
     }
 
-    pub(crate) fn get_kernel_bytes_xlated(&self) -> Result<u32, BpfmanError> {
+    pub fn get_kernel_bytes_xlated(&self) -> Result<u32, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_BYTES_XLATED).map(bytes_to_u32)
     }
 
@@ -862,7 +714,7 @@ impl ProgramData {
         )
     }
 
-    pub(crate) fn get_kernel_jited(&self) -> Result<bool, BpfmanError> {
+    pub fn get_kernel_jited(&self) -> Result<bool, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_JITED).map(bytes_to_bool)
     }
 
@@ -874,7 +726,7 @@ impl ProgramData {
         )
     }
 
-    pub(crate) fn get_kernel_bytes_jited(&self) -> Result<u32, BpfmanError> {
+    pub fn get_kernel_bytes_jited(&self) -> Result<u32, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_BYTES_JITED).map(bytes_to_u32)
     }
 
@@ -886,7 +738,7 @@ impl ProgramData {
         )
     }
 
-    pub(crate) fn get_kernel_bytes_memlock(&self) -> Result<u32, BpfmanError> {
+    pub fn get_kernel_bytes_memlock(&self) -> Result<u32, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_BYTES_MEMLOCK).map(bytes_to_u32)
     }
 
@@ -901,7 +753,7 @@ impl ProgramData {
         )
     }
 
-    pub(crate) fn get_kernel_verified_insns(&self) -> Result<u32, BpfmanError> {
+    pub fn get_kernel_verified_insns(&self) -> Result<u32, BpfmanError> {
         sled_get(&self.db_tree, KERNEL_VERIFIED_INSNS).map(bytes_to_u32)
     }
 
@@ -975,7 +827,7 @@ impl XdpProgram {
         sled_insert(&self.data.db_tree, XDP_PRIORITY, &priority.to_ne_bytes())
     }
 
-    pub(crate) fn get_priority(&self) -> Result<i32, BpfmanError> {
+    pub fn get_priority(&self) -> Result<i32, BpfmanError> {
         sled_get(&self.data.db_tree, XDP_PRIORITY).map(bytes_to_i32)
     }
 
@@ -983,7 +835,7 @@ impl XdpProgram {
         sled_insert(&self.data.db_tree, XDP_IFACE, iface.as_bytes())
     }
 
-    pub(crate) fn get_iface(&self) -> Result<String, BpfmanError> {
+    pub fn get_iface(&self) -> Result<String, BpfmanError> {
         sled_get(&self.data.db_tree, XDP_IFACE).map(|v| bytes_to_string(&v))
     }
 
@@ -1001,7 +853,7 @@ impl XdpProgram {
             })
     }
 
-    pub(crate) fn get_proceed_on(&self) -> Result<XdpProceedOn, BpfmanError> {
+    pub fn get_proceed_on(&self) -> Result<XdpProceedOn, BpfmanError> {
         self.data
             .db_tree
             .scan_prefix(PREFIX_XDP_PROCEED_ON)
@@ -1024,7 +876,7 @@ impl XdpProgram {
         sled_insert(&self.data.db_tree, XDP_CURRENT_POSITION, &pos.to_ne_bytes())
     }
 
-    pub(crate) fn get_current_position(&self) -> Result<Option<usize>, BpfmanError> {
+    pub fn get_current_position(&self) -> Result<Option<usize>, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, XDP_CURRENT_POSITION)?.map(bytes_to_usize))
     }
 
@@ -1032,7 +884,7 @@ impl XdpProgram {
         sled_insert(&self.data.db_tree, XDP_IF_INDEX, &if_index.to_ne_bytes())
     }
 
-    pub(crate) fn get_if_index(&self) -> Result<Option<u32>, BpfmanError> {
+    pub fn get_if_index(&self) -> Result<Option<u32>, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, XDP_IF_INDEX)?.map(bytes_to_u32))
     }
 
@@ -1044,7 +896,7 @@ impl XdpProgram {
         )
     }
 
-    pub(crate) fn get_attached(&self) -> Result<bool, BpfmanError> {
+    pub fn get_attached(&self) -> Result<bool, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, XDP_ATTACHED)?
             .map(bytes_to_bool)
             .unwrap_or(false))
@@ -1087,7 +939,7 @@ impl TcProgram {
         sled_insert(&self.data.db_tree, TC_PRIORITY, &priority.to_ne_bytes())
     }
 
-    pub(crate) fn get_priority(&self) -> Result<i32, BpfmanError> {
+    pub fn get_priority(&self) -> Result<i32, BpfmanError> {
         sled_get(&self.data.db_tree, TC_PRIORITY).map(bytes_to_i32)
     }
 
@@ -1095,7 +947,7 @@ impl TcProgram {
         sled_insert(&self.data.db_tree, TC_IFACE, iface.as_bytes())
     }
 
-    pub(crate) fn get_iface(&self) -> Result<String, BpfmanError> {
+    pub fn get_iface(&self) -> Result<String, BpfmanError> {
         sled_get(&self.data.db_tree, TC_IFACE).map(|v| bytes_to_string(&v))
     }
 
@@ -1113,7 +965,7 @@ impl TcProgram {
             })
     }
 
-    pub(crate) fn get_proceed_on(&self) -> Result<TcProceedOn, BpfmanError> {
+    pub fn get_proceed_on(&self) -> Result<TcProceedOn, BpfmanError> {
         self.data
             .db_tree
             .scan_prefix(PREFIX_TC_PROCEED_ON)
@@ -1133,7 +985,7 @@ impl TcProgram {
         sled_insert(&self.data.db_tree, TC_CURRENT_POSITION, &pos.to_ne_bytes())
     }
 
-    pub(crate) fn get_current_position(&self) -> Result<Option<usize>, BpfmanError> {
+    pub fn get_current_position(&self) -> Result<Option<usize>, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, TC_CURRENT_POSITION)?.map(bytes_to_usize))
     }
 
@@ -1141,7 +993,7 @@ impl TcProgram {
         sled_insert(&self.data.db_tree, TC_IF_INDEX, &if_index.to_ne_bytes())
     }
 
-    pub(crate) fn get_if_index(&self) -> Result<Option<u32>, BpfmanError> {
+    pub fn get_if_index(&self) -> Result<Option<u32>, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, TC_IF_INDEX)?.map(bytes_to_u32))
     }
 
@@ -1153,7 +1005,7 @@ impl TcProgram {
         )
     }
 
-    pub(crate) fn get_attached(&self) -> Result<bool, BpfmanError> {
+    pub fn get_attached(&self) -> Result<bool, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, TC_ATTACHED)?
             .map(bytes_to_bool)
             .unwrap_or(false))
@@ -1167,7 +1019,7 @@ impl TcProgram {
         )
     }
 
-    pub(crate) fn get_direction(&self) -> Result<Direction, BpfmanError> {
+    pub fn get_direction(&self) -> Result<Direction, BpfmanError> {
         sled_get(&self.data.db_tree, TC_DIRECTION)
             .map(|v| bytes_to_string(&v).to_string().try_into().unwrap())
     }
@@ -1199,7 +1051,7 @@ impl TracepointProgram {
         sled_insert(&self.data.db_tree, TRACEPOINT_NAME, tracepoint.as_bytes())
     }
 
-    pub(crate) fn get_tracepoint(&self) -> Result<String, BpfmanError> {
+    pub fn get_tracepoint(&self) -> Result<String, BpfmanError> {
         sled_get(&self.data.db_tree, TRACEPOINT_NAME).map(|v| bytes_to_string(&v))
     }
 
@@ -1240,7 +1092,7 @@ impl KprobeProgram {
         sled_insert(&self.data.db_tree, KPROBE_FN_NAME, fn_name.as_bytes())
     }
 
-    pub(crate) fn get_fn_name(&self) -> Result<String, BpfmanError> {
+    pub fn get_fn_name(&self) -> Result<String, BpfmanError> {
         sled_get(&self.data.db_tree, KPROBE_FN_NAME).map(|v| bytes_to_string(&v))
     }
 
@@ -1248,7 +1100,7 @@ impl KprobeProgram {
         sled_insert(&self.data.db_tree, KPROBE_OFFSET, &offset.to_ne_bytes())
     }
 
-    pub(crate) fn get_offset(&self) -> Result<u64, BpfmanError> {
+    pub fn get_offset(&self) -> Result<u64, BpfmanError> {
         sled_get(&self.data.db_tree, KPROBE_OFFSET).map(bytes_to_u64)
     }
 
@@ -1260,7 +1112,7 @@ impl KprobeProgram {
         )
     }
 
-    pub(crate) fn get_retprobe(&self) -> Result<bool, BpfmanError> {
+    pub fn get_retprobe(&self) -> Result<bool, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, KPROBE_RETPROBE)?
             .map(bytes_to_bool)
             .unwrap_or(false))
@@ -1274,7 +1126,7 @@ impl KprobeProgram {
         )
     }
 
-    pub(crate) fn get_container_pid(&self) -> Result<Option<i32>, BpfmanError> {
+    pub fn get_container_pid(&self) -> Result<Option<i32>, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, KPROBE_CONTAINER_PID)?.map(bytes_to_i32))
     }
 
@@ -1325,7 +1177,7 @@ impl UprobeProgram {
         sled_insert(&self.data.db_tree, UPROBE_FN_NAME, fn_name.as_bytes())
     }
 
-    pub(crate) fn get_fn_name(&self) -> Result<Option<String>, BpfmanError> {
+    pub fn get_fn_name(&self) -> Result<Option<String>, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, UPROBE_FN_NAME)?.map(|v| bytes_to_string(&v)))
     }
 
@@ -1333,7 +1185,7 @@ impl UprobeProgram {
         sled_insert(&self.data.db_tree, UPROBE_OFFSET, &offset.to_ne_bytes())
     }
 
-    pub(crate) fn get_offset(&self) -> Result<u64, BpfmanError> {
+    pub fn get_offset(&self) -> Result<u64, BpfmanError> {
         sled_get(&self.data.db_tree, UPROBE_OFFSET).map(bytes_to_u64)
     }
 
@@ -1345,7 +1197,7 @@ impl UprobeProgram {
         )
     }
 
-    pub(crate) fn get_retprobe(&self) -> Result<bool, BpfmanError> {
+    pub fn get_retprobe(&self) -> Result<bool, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, UPROBE_RETPROBE)?
             .map(bytes_to_bool)
             .unwrap_or(false))
@@ -1359,7 +1211,7 @@ impl UprobeProgram {
         )
     }
 
-    pub(crate) fn get_container_pid(&self) -> Result<Option<i32>, BpfmanError> {
+    pub fn get_container_pid(&self) -> Result<Option<i32>, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, UPROBE_CONTAINER_PID)?.map(bytes_to_i32))
     }
 
@@ -1367,7 +1219,7 @@ impl UprobeProgram {
         sled_insert(&self.data.db_tree, UPROBE_PID, &pid.to_ne_bytes())
     }
 
-    pub(crate) fn get_pid(&self) -> Result<Option<i32>, BpfmanError> {
+    pub fn get_pid(&self) -> Result<Option<i32>, BpfmanError> {
         Ok(sled_get_option(&self.data.db_tree, UPROBE_PID)?.map(bytes_to_i32))
     }
 
@@ -1375,7 +1227,7 @@ impl UprobeProgram {
         sled_insert(&self.data.db_tree, UPROBE_TARGET, target.as_bytes())
     }
 
-    pub(crate) fn get_target(&self) -> Result<String, BpfmanError> {
+    pub fn get_target(&self) -> Result<String, BpfmanError> {
         sled_get(&self.data.db_tree, UPROBE_TARGET).map(|v| bytes_to_string(&v))
     }
 
@@ -1406,7 +1258,7 @@ impl FentryProgram {
         sled_insert(&self.data.db_tree, FENTRY_FN_NAME, fn_name.as_bytes())
     }
 
-    pub(crate) fn get_fn_name(&self) -> Result<String, BpfmanError> {
+    pub fn get_fn_name(&self) -> Result<String, BpfmanError> {
         sled_get(&self.data.db_tree, FENTRY_FN_NAME).map(|v| bytes_to_string(&v))
     }
 
@@ -1437,7 +1289,7 @@ impl FexitProgram {
         sled_insert(&self.data.db_tree, FEXIT_FN_NAME, fn_name.as_bytes())
     }
 
-    pub(crate) fn get_fn_name(&self) -> Result<String, BpfmanError> {
+    pub fn get_fn_name(&self) -> Result<String, BpfmanError> {
         sled_get(&self.data.db_tree, FEXIT_FN_NAME).map(|v| bytes_to_string(&v))
     }
 
@@ -1451,7 +1303,7 @@ impl FexitProgram {
 }
 
 impl Program {
-    pub(crate) fn kind(&self) -> ProgramType {
+    pub fn kind(&self) -> ProgramType {
         match self {
             Program::Xdp(_) => ProgramType::Xdp,
             Program::Tc(_) => ProgramType::Tc,
@@ -1519,9 +1371,9 @@ impl Program {
         }
     }
 
-    pub fn delete(&self) -> Result<(), anyhow::Error> {
+    pub(crate) fn delete(&self, root_db: &Db) -> Result<(), anyhow::Error> {
         let id = self.get_data().get_id()?;
-        ROOT_DB.drop_tree(self.get_data().db_tree.name())?;
+        root_db.drop_tree(self.get_data().db_tree.name())?;
 
         let path = format!("{RTDIR_FS}/prog_{id}");
         if PathBuf::from(&path).exists() {
@@ -1574,21 +1426,6 @@ impl Program {
         }
     }
 
-    pub(crate) fn location(&self) -> Result<Location, BpfmanError> {
-        match self {
-            Program::Xdp(p) => p.data.get_location(),
-            Program::Tracepoint(p) => p.data.get_location(),
-            Program::Tc(p) => p.data.get_location(),
-            Program::Kprobe(p) => p.data.get_location(),
-            Program::Uprobe(p) => p.data.get_location(),
-            Program::Fentry(p) => p.data.get_location(),
-            Program::Fexit(p) => p.data.get_location(),
-            Program::Unsupported(_) => Err(BpfmanError::Error(
-                "cannot get location for unsupported programs".to_string(),
-            )),
-        }
-    }
-
     pub(crate) fn direction(&self) -> Result<Option<Direction>, BpfmanError> {
         match self {
             Program::Tc(p) => Ok(Some(p.get_direction()?)),
@@ -1596,7 +1433,7 @@ impl Program {
         }
     }
 
-    pub(crate) fn get_data(&self) -> &ProgramData {
+    pub fn get_data(&self) -> &ProgramData {
         match self {
             Program::Xdp(p) => p.get_data(),
             Program::Tracepoint(p) => p.get_data(),
@@ -1641,6 +1478,640 @@ impl Program {
                 _ => Err(BpfmanError::Error("Unsupported program type".to_string())),
             },
             None => Err(BpfmanError::Error("Unsupported program type".to_string())),
+        }
+    }
+}
+
+#[derive(ValueEnum, Copy, Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
+pub enum ProgramType {
+    Unspec,
+    SocketFilter,
+    Probe, // kprobe, kretprobe, uprobe, uretprobe
+    Tc,
+    SchedAct,
+    Tracepoint,
+    Xdp,
+    PerfEvent,
+    CgroupSkb,
+    CgroupSock,
+    LwtIn,
+    LwtOut,
+    LwtXmit,
+    SockOps,
+    SkSkb,
+    CgroupDevice,
+    SkMsg,
+    RawTracepoint,
+    CgroupSockAddr,
+    LwtSeg6Local,
+    LircMode2,
+    SkReuseport,
+    FlowDissector,
+    CgroupSysctl,
+    RawTracepointWritable,
+    CgroupSockopt,
+    Tracing, // fentry, fexit
+    StructOps,
+    Ext,
+    Lsm,
+    SkLookup,
+    Syscall,
+}
+
+impl TryFrom<String> for ProgramType {
+    type Error = ParseError;
+
+    fn try_from(value: String) -> Result<Self, Self::Error> {
+        Ok(match value.as_str() {
+            "unspec" => ProgramType::Unspec,
+            "socket_filter" => ProgramType::SocketFilter,
+            "probe" => ProgramType::Probe,
+            "tc" => ProgramType::Tc,
+            "sched_act" => ProgramType::SchedAct,
+            "tracepoint" => ProgramType::Tracepoint,
+            "xdp" => ProgramType::Xdp,
+            "perf_event" => ProgramType::PerfEvent,
+            "cgroup_skb" => ProgramType::CgroupSkb,
+            "cgroup_sock" => ProgramType::CgroupSock,
+            "lwt_in" => ProgramType::LwtIn,
+            "lwt_out" => ProgramType::LwtOut,
+            "lwt_xmit" => ProgramType::LwtXmit,
+            "sock_ops" => ProgramType::SockOps,
+            "sk_skb" => ProgramType::SkSkb,
+            "cgroup_device" => ProgramType::CgroupDevice,
+            "sk_msg" => ProgramType::SkMsg,
+            "raw_tracepoint" => ProgramType::RawTracepoint,
+            "cgroup_sock_addr" => ProgramType::CgroupSockAddr,
+            "lwt_seg6local" => ProgramType::LwtSeg6Local,
+            "lirc_mode2" => ProgramType::LircMode2,
+            "sk_reuseport" => ProgramType::SkReuseport,
+            "flow_dissector" => ProgramType::FlowDissector,
+            "cgroup_sysctl" => ProgramType::CgroupSysctl,
+            "raw_tracepoint_writable" => ProgramType::RawTracepointWritable,
+            "cgroup_sockopt" => ProgramType::CgroupSockopt,
+            "tracing" => ProgramType::Tracing,
+            "struct_ops" => ProgramType::StructOps,
+            "ext" => ProgramType::Ext,
+            "lsm" => ProgramType::Lsm,
+            "sk_lookup" => ProgramType::SkLookup,
+            "syscall" => ProgramType::Syscall,
+            other => {
+                return Err(ParseError::InvalidProgramType {
+                    program: other.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl TryFrom<u32> for ProgramType {
+    type Error = ParseError;
+
+    fn try_from(value: u32) -> Result<Self, Self::Error> {
+        Ok(match value {
+            0 => ProgramType::Unspec,
+            1 => ProgramType::SocketFilter,
+            2 => ProgramType::Probe,
+            3 => ProgramType::Tc,
+            4 => ProgramType::SchedAct,
+            5 => ProgramType::Tracepoint,
+            6 => ProgramType::Xdp,
+            7 => ProgramType::PerfEvent,
+            8 => ProgramType::CgroupSkb,
+            9 => ProgramType::CgroupSock,
+            10 => ProgramType::LwtIn,
+            11 => ProgramType::LwtOut,
+            12 => ProgramType::LwtXmit,
+            13 => ProgramType::SockOps,
+            14 => ProgramType::SkSkb,
+            15 => ProgramType::CgroupDevice,
+            16 => ProgramType::SkMsg,
+            17 => ProgramType::RawTracepoint,
+            18 => ProgramType::CgroupSockAddr,
+            19 => ProgramType::LwtSeg6Local,
+            20 => ProgramType::LircMode2,
+            21 => ProgramType::SkReuseport,
+            22 => ProgramType::FlowDissector,
+            23 => ProgramType::CgroupSysctl,
+            24 => ProgramType::RawTracepointWritable,
+            25 => ProgramType::CgroupSockopt,
+            26 => ProgramType::Tracing,
+            27 => ProgramType::StructOps,
+            28 => ProgramType::Ext,
+            29 => ProgramType::Lsm,
+            30 => ProgramType::SkLookup,
+            31 => ProgramType::Syscall,
+            other => {
+                return Err(ParseError::InvalidProgramType {
+                    program: other.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl From<ProgramType> for u32 {
+    fn from(val: ProgramType) -> Self {
+        match val {
+            ProgramType::Unspec => 0,
+            ProgramType::SocketFilter => 1,
+            ProgramType::Probe => 2,
+            ProgramType::Tc => 3,
+            ProgramType::SchedAct => 4,
+            ProgramType::Tracepoint => 5,
+            ProgramType::Xdp => 6,
+            ProgramType::PerfEvent => 7,
+            ProgramType::CgroupSkb => 8,
+            ProgramType::CgroupSock => 9,
+            ProgramType::LwtIn => 10,
+            ProgramType::LwtOut => 11,
+            ProgramType::LwtXmit => 12,
+            ProgramType::SockOps => 13,
+            ProgramType::SkSkb => 14,
+            ProgramType::CgroupDevice => 15,
+            ProgramType::SkMsg => 16,
+            ProgramType::RawTracepoint => 17,
+            ProgramType::CgroupSockAddr => 18,
+            ProgramType::LwtSeg6Local => 19,
+            ProgramType::LircMode2 => 20,
+            ProgramType::SkReuseport => 21,
+            ProgramType::FlowDissector => 22,
+            ProgramType::CgroupSysctl => 23,
+            ProgramType::RawTracepointWritable => 24,
+            ProgramType::CgroupSockopt => 25,
+            ProgramType::Tracing => 26,
+            ProgramType::StructOps => 27,
+            ProgramType::Ext => 28,
+            ProgramType::Lsm => 29,
+            ProgramType::SkLookup => 30,
+            ProgramType::Syscall => 31,
+        }
+    }
+}
+
+impl std::fmt::Display for ProgramType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let v = match self {
+            ProgramType::Unspec => "unspec",
+            ProgramType::SocketFilter => "socket_filter",
+            ProgramType::Probe => "probe",
+            ProgramType::Tc => "tc",
+            ProgramType::SchedAct => "sched_act",
+            ProgramType::Tracepoint => "tracepoint",
+            ProgramType::Xdp => "xdp",
+            ProgramType::PerfEvent => "perf_event",
+            ProgramType::CgroupSkb => "cgroup_skb",
+            ProgramType::CgroupSock => "cgroup_sock",
+            ProgramType::LwtIn => "lwt_in",
+            ProgramType::LwtOut => "lwt_out",
+            ProgramType::LwtXmit => "lwt_xmit",
+            ProgramType::SockOps => "sock_ops",
+            ProgramType::SkSkb => "sk_skb",
+            ProgramType::CgroupDevice => "cgroup_device",
+            ProgramType::SkMsg => "sk_msg",
+            ProgramType::RawTracepoint => "raw_tracepoint",
+            ProgramType::CgroupSockAddr => "cgroup_sock_addr",
+            ProgramType::LwtSeg6Local => "lwt_seg6local",
+            ProgramType::LircMode2 => "lirc_mode2",
+            ProgramType::SkReuseport => "sk_reuseport",
+            ProgramType::FlowDissector => "flow_dissector",
+            ProgramType::CgroupSysctl => "cgroup_sysctl",
+            ProgramType::RawTracepointWritable => "raw_tracepoint_writable",
+            ProgramType::CgroupSockopt => "cgroup_sockopt",
+            ProgramType::Tracing => "tracing",
+            ProgramType::StructOps => "struct_ops",
+            ProgramType::Ext => "ext",
+            ProgramType::Lsm => "lsm",
+            ProgramType::SkLookup => "sk_lookup",
+            ProgramType::Syscall => "syscall",
+        };
+        write!(f, "{v}")
+    }
+}
+
+#[derive(Copy, Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
+pub enum ProbeType {
+    Kprobe,
+    Kretprobe,
+    Uprobe,
+    Uretprobe,
+}
+
+impl TryFrom<i32> for ProbeType {
+    type Error = ParseError;
+
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        Ok(match value {
+            0 => ProbeType::Kprobe,
+            1 => ProbeType::Kretprobe,
+            2 => ProbeType::Uprobe,
+            3 => ProbeType::Uretprobe,
+            other => {
+                return Err(ParseError::InvalidProbeType {
+                    probe: other.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl From<aya::programs::ProbeKind> for ProbeType {
+    fn from(value: aya::programs::ProbeKind) -> Self {
+        match value {
+            aya::programs::ProbeKind::KProbe => ProbeType::Kprobe,
+            aya::programs::ProbeKind::KRetProbe => ProbeType::Kretprobe,
+            aya::programs::ProbeKind::UProbe => ProbeType::Uprobe,
+            aya::programs::ProbeKind::URetProbe => ProbeType::Uretprobe,
+        }
+    }
+}
+
+impl std::fmt::Display for ProbeType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let v = match self {
+            ProbeType::Kprobe => "kprobe",
+            ProbeType::Kretprobe => "kretprobe",
+            ProbeType::Uprobe => "uprobe",
+            ProbeType::Uretprobe => "uretprobe",
+        };
+        write!(f, "{v}")
+    }
+}
+
+#[derive(Serialize, Deserialize, Copy, Clone, Debug)]
+pub enum XdpProceedOnEntry {
+    Aborted,
+    Drop,
+    Pass,
+    Tx,
+    Redirect,
+    DispatcherReturn = 31,
+}
+
+impl FromIterator<XdpProceedOnEntry> for XdpProceedOn {
+    fn from_iter<I: IntoIterator<Item = XdpProceedOnEntry>>(iter: I) -> Self {
+        let mut c = Vec::new();
+
+        let mut iter = iter.into_iter().peekable();
+
+        // make sure to default if proceed on is empty
+        if iter.peek().is_none() {
+            return XdpProceedOn::default();
+        };
+
+        for i in iter {
+            c.push(i);
+        }
+
+        XdpProceedOn(c)
+    }
+}
+
+impl TryFrom<String> for XdpProceedOnEntry {
+    type Error = ParseError;
+    fn try_from(value: String) -> Result<Self, Self::Error> {
+        Ok(match value.as_str() {
+            "aborted" => XdpProceedOnEntry::Aborted,
+            "drop" => XdpProceedOnEntry::Drop,
+            "pass" => XdpProceedOnEntry::Pass,
+            "tx" => XdpProceedOnEntry::Tx,
+            "redirect" => XdpProceedOnEntry::Redirect,
+            "dispatcher_return" => XdpProceedOnEntry::DispatcherReturn,
+            proceedon => {
+                return Err(ParseError::InvalidProceedOn {
+                    proceedon: proceedon.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl TryFrom<i32> for XdpProceedOnEntry {
+    type Error = ParseError;
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        Ok(match value {
+            0 => XdpProceedOnEntry::Aborted,
+            1 => XdpProceedOnEntry::Drop,
+            2 => XdpProceedOnEntry::Pass,
+            3 => XdpProceedOnEntry::Tx,
+            4 => XdpProceedOnEntry::Redirect,
+            31 => XdpProceedOnEntry::DispatcherReturn,
+            proceedon => {
+                return Err(ParseError::InvalidProceedOn {
+                    proceedon: proceedon.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl std::fmt::Display for XdpProceedOnEntry {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let v = match self {
+            XdpProceedOnEntry::Aborted => "aborted",
+            XdpProceedOnEntry::Drop => "drop",
+            XdpProceedOnEntry::Pass => "pass",
+            XdpProceedOnEntry::Tx => "tx",
+            XdpProceedOnEntry::Redirect => "redirect",
+            XdpProceedOnEntry::DispatcherReturn => "dispatcher_return",
+        };
+        write!(f, "{v}")
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct XdpProceedOn(Vec<XdpProceedOnEntry>);
+impl Default for XdpProceedOn {
+    fn default() -> Self {
+        XdpProceedOn(vec![
+            XdpProceedOnEntry::Pass,
+            XdpProceedOnEntry::DispatcherReturn,
+        ])
+    }
+}
+
+impl XdpProceedOn {
+    pub fn from_strings<T: AsRef<[String]>>(values: T) -> Result<XdpProceedOn, ParseError> {
+        let entries = values.as_ref();
+        let mut res = vec![];
+        for e in entries {
+            res.push(e.to_owned().try_into()?)
+        }
+        Ok(XdpProceedOn(res))
+    }
+
+    pub fn from_int32s<T: AsRef<[i32]>>(values: T) -> Result<XdpProceedOn, ParseError> {
+        let entries = values.as_ref();
+        if entries.is_empty() {
+            return Ok(XdpProceedOn::default());
+        }
+        let mut res = vec![];
+        for e in entries {
+            res.push((*e).try_into()?)
+        }
+        Ok(XdpProceedOn(res))
+    }
+
+    pub fn mask(&self) -> u32 {
+        let mut proceed_on_mask: u32 = 0;
+        for action in self.0.clone().into_iter() {
+            proceed_on_mask |= 1 << action as u32;
+        }
+        proceed_on_mask
+    }
+
+    pub fn as_action_vec(&self) -> Vec<i32> {
+        let mut res = vec![];
+        for entry in &self.0 {
+            res.push((*entry) as i32)
+        }
+        res
+    }
+}
+
+impl std::fmt::Display for XdpProceedOn {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let res: Vec<String> = self.0.iter().map(|x| x.to_string()).collect();
+        write!(f, "{}", res.join(", "))
+    }
+}
+
+#[derive(Serialize, Deserialize, Copy, Clone, Debug)]
+pub enum TcProceedOnEntry {
+    Unspec = -1,
+    Ok = 0,
+    Reclassify,
+    Shot,
+    Pipe,
+    Stolen,
+    Queued,
+    Repeat,
+    Redirect,
+    Trap,
+    DispatcherReturn = 30,
+}
+
+impl TryFrom<String> for TcProceedOnEntry {
+    type Error = ParseError;
+    fn try_from(value: String) -> Result<Self, Self::Error> {
+        Ok(match value.as_str() {
+            "unspec" => TcProceedOnEntry::Unspec,
+            "ok" => TcProceedOnEntry::Ok,
+            "reclassify" => TcProceedOnEntry::Reclassify,
+            "shot" => TcProceedOnEntry::Shot,
+            "pipe" => TcProceedOnEntry::Pipe,
+            "stolen" => TcProceedOnEntry::Stolen,
+            "queued" => TcProceedOnEntry::Queued,
+            "repeat" => TcProceedOnEntry::Repeat,
+            "redirect" => TcProceedOnEntry::Redirect,
+            "trap" => TcProceedOnEntry::Trap,
+            "dispatcher_return" => TcProceedOnEntry::DispatcherReturn,
+            proceedon => {
+                return Err(ParseError::InvalidProceedOn {
+                    proceedon: proceedon.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl TryFrom<i32> for TcProceedOnEntry {
+    type Error = ParseError;
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        Ok(match value {
+            -1 => TcProceedOnEntry::Unspec,
+            0 => TcProceedOnEntry::Ok,
+            1 => TcProceedOnEntry::Reclassify,
+            2 => TcProceedOnEntry::Shot,
+            3 => TcProceedOnEntry::Pipe,
+            4 => TcProceedOnEntry::Stolen,
+            5 => TcProceedOnEntry::Queued,
+            6 => TcProceedOnEntry::Repeat,
+            7 => TcProceedOnEntry::Redirect,
+            8 => TcProceedOnEntry::Trap,
+            30 => TcProceedOnEntry::DispatcherReturn,
+            proceedon => {
+                return Err(ParseError::InvalidProceedOn {
+                    proceedon: proceedon.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl std::fmt::Display for TcProceedOnEntry {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let v = match self {
+            TcProceedOnEntry::Unspec => "unspec",
+            TcProceedOnEntry::Ok => "ok",
+            TcProceedOnEntry::Reclassify => "reclassify",
+            TcProceedOnEntry::Shot => "shot",
+            TcProceedOnEntry::Pipe => "pipe",
+            TcProceedOnEntry::Stolen => "stolen",
+            TcProceedOnEntry::Queued => "queued",
+            TcProceedOnEntry::Repeat => "repeat",
+            TcProceedOnEntry::Redirect => "redirect",
+            TcProceedOnEntry::Trap => "trap",
+            TcProceedOnEntry::DispatcherReturn => "dispatcher_return",
+        };
+        write!(f, "{v}")
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct TcProceedOn(pub(crate) Vec<TcProceedOnEntry>);
+impl Default for TcProceedOn {
+    fn default() -> Self {
+        TcProceedOn(vec![
+            TcProceedOnEntry::Pipe,
+            TcProceedOnEntry::DispatcherReturn,
+        ])
+    }
+}
+
+impl FromIterator<TcProceedOnEntry> for TcProceedOn {
+    fn from_iter<I: IntoIterator<Item = TcProceedOnEntry>>(iter: I) -> Self {
+        let mut c = Vec::new();
+        let mut iter = iter.into_iter().peekable();
+
+        // make sure to default if proceed on is empty
+        if iter.peek().is_none() {
+            return TcProceedOn::default();
+        };
+
+        for i in iter {
+            c.push(i);
+        }
+
+        TcProceedOn(c)
+    }
+}
+
+impl TcProceedOn {
+    pub fn from_strings<T: AsRef<[String]>>(values: T) -> Result<TcProceedOn, ParseError> {
+        let entries = values.as_ref();
+        let mut res = vec![];
+        for e in entries {
+            res.push(e.to_owned().try_into()?)
+        }
+        Ok(TcProceedOn(res))
+    }
+
+    pub fn from_int32s<T: AsRef<[i32]>>(values: T) -> Result<TcProceedOn, ParseError> {
+        let entries = values.as_ref();
+        if entries.is_empty() {
+            return Ok(TcProceedOn::default());
+        }
+        let mut res = vec![];
+        for e in entries {
+            res.push((*e).try_into()?)
+        }
+        Ok(TcProceedOn(res))
+    }
+
+    // Valid TC return values range from -1 to 8.  Since -1 is not a valid shift value,
+    // 1 is added to the value to determine the bit to set in the bitmask and,
+    // correspondingly, The TC dispatcher adds 1 to the return value from the BPF program
+    // before it compares it to the configured bit mask.
+    pub fn mask(&self) -> u32 {
+        let mut proceed_on_mask: u32 = 0;
+        for action in self.0.clone().into_iter() {
+            proceed_on_mask |= 1 << ((action as i32) + 1);
+        }
+        proceed_on_mask
+    }
+
+    pub fn as_action_vec(&self) -> Vec<i32> {
+        let mut res = vec![];
+        for entry in &self.0 {
+            res.push((*entry) as i32)
+        }
+        res
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
+}
+
+impl std::fmt::Display for TcProceedOn {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let res: Vec<String> = self.0.iter().map(|x| x.to_string()).collect();
+        write!(f, "{}", res.join(", "))
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub enum ImagePullPolicy {
+    Always,
+    IfNotPresent,
+    Never,
+}
+
+impl std::fmt::Display for ImagePullPolicy {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let v = match self {
+            ImagePullPolicy::Always => "Always",
+            ImagePullPolicy::IfNotPresent => "IfNotPresent",
+            ImagePullPolicy::Never => "Never",
+        };
+        write!(f, "{v}")
+    }
+}
+
+impl TryFrom<i32> for ImagePullPolicy {
+    type Error = ParseError;
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        Ok(match value {
+            0 => ImagePullPolicy::Always,
+            1 => ImagePullPolicy::IfNotPresent,
+            2 => ImagePullPolicy::Never,
+            policy => {
+                return Err(ParseError::InvalidBytecodeImagePullPolicy {
+                    pull_policy: policy.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl TryFrom<&str> for ImagePullPolicy {
+    type Error = ParseError;
+    fn try_from(value: &str) -> Result<Self, Self::Error> {
+        Ok(match value {
+            "Always" => ImagePullPolicy::Always,
+            "IfNotPresent" => ImagePullPolicy::IfNotPresent,
+            "Never" => ImagePullPolicy::Never,
+            policy => {
+                return Err(ParseError::InvalidBytecodeImagePullPolicy {
+                    pull_policy: policy.to_string(),
+                })
+            }
+        })
+    }
+}
+
+impl From<ImagePullPolicy> for i32 {
+    fn from(value: ImagePullPolicy) -> Self {
+        match value {
+            ImagePullPolicy::Always => 0,
+            ImagePullPolicy::IfNotPresent => 1,
+            ImagePullPolicy::Never => 2,
+        }
+    }
+}
+
+impl std::fmt::Display for Location {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match &self {
+            // Cast imagePullPolicy into it's concrete type so we can easily print.
+            Location::Image(i) => write!(
+                f,
+                "image: {{ url: {}, pullpolicy: {} }}",
+                i.image_url,
+                TryInto::<ImagePullPolicy>::try_into(i.image_pull_policy.clone()).unwrap()
+            ),
+            Location::File(p) => write!(f, "file: {{ path: {p} }}"),
         }
     }
 }
