+++ tests/integration-test/Cargo.toml
@@ -7,7 +7,7 @@ version = "0.1.0"
 [dependencies]
 anyhow = { workspace = true, features = ["std"] }
 assert_cmd = { workspace = true }
-bpfman-api = { workspace = true }
+bpfman = { workspace = true }
 env_logger = { workspace = true }
 integration-test-macros = { workspace = true }
 inventory = { workspace = true }
@@ -15,4 +15,3 @@ log = { workspace = true }
 predicates = { workspace = true }
 rand = { workspace = true, features = ["std", "std_rng"] }
 regex = { workspace = true, features = ["unicode-perl"] }
-sled = { workspace = true }
