+++ tests/integration-test/src/tests/e2e.rs
@@ -1,6 +1,6 @@
 use std::{path::PathBuf, thread::sleep, time::Duration};
 
-use bpfman_api::util::directories::{RTDIR_FS_TC_EGRESS, RTDIR_FS_TC_INGRESS, RTDIR_FS_XDP};
+use bpfman::directories::{RTDIR_FS_TC_EGRESS, RTDIR_FS_TC_INGRESS, RTDIR_FS_XDP};
 use log::debug;
 
 use super::{integration_test, IntegrationTest};
